@@ -4,9 +4,7 @@ import JJorm from "../JJorm";
 import Bind from "../ReactBootstrap";
 import L from "../@global/Language";
 
-export default class Index extends JJorm<JJWAK.Page.Props>{
-  ACTION_RECEIVER_TABLE = {};
-
+export default class Index extends JJorm<JJWAK.Page.Props<'Index'>>{
   render():React.ReactNode{
     return <article>
       Hello, {this.props.page}
